Break up code, tests separate from core logic
