+++ Source/CesiumRuntime/Private/Tests/CesiumLoadTestCore.cpp
@@ -2,34 +2,16 @@
 
 #if WITH_EDITOR
 
-#include "Framework/Application/SlateApplication.h"
-#include "Misc/AutomationTest.h"
+#include "CesiumLoadTestCore.h"
+
+#include "CesiumAsync/ICacheDatabase.h"
+#include "CesiumRuntime.h"
+
+#include "Editor.h"
 #include "Tests/AutomationCommon.h"
 #include "Tests/AutomationEditorCommon.h"
 
-#include "CesiumAsync/ICacheDatabase.h"
-#include "CesiumGltfComponent.h"
-#include "CesiumRuntime.h"
-#include "CesiumSceneGeneration.h"
-#include "CesiumTestHelpers.h"
-#include "GlobeAwareDefaultPawn.h"
-
-using namespace Cesium;
-
-IMPLEMENT_SIMPLE_AUTOMATION_TEST(
-    FCesiumLoadTestDenver,
-    "Cesium.Performance.LoadTestDenver",
-    EAutomationTestFlags::EditorContext | EAutomationTestFlags::PerfFilter)
-
-IMPLEMENT_SIMPLE_AUTOMATION_TEST(
-    FCesiumLoadTestGoogleplex,
-    "Cesium.Performance.LoadTestGoogleplex",
-    EAutomationTestFlags::EditorContext | EAutomationTestFlags::PerfFilter)
-
-IMPLEMENT_SIMPLE_AUTOMATION_TEST(
-    FCesiumLoadTestMontrealPointCloud,
-    "Cesium.Performance.LoadTestMontrealPointCloud",
-    EAutomationTestFlags::EditorContext | EAutomationTestFlags::PerfFilter)
+namespace Cesium {
 
 struct LoadTestContext {
   SceneGenerationContext creationContext;
@@ -143,12 +125,6 @@ bool WaitForPIECommand::Update() {
   return true;
 }
 
-struct TestPass {
-  FString name;
-  std::function<void(SceneGenerationContext&)> setupStep;
-  std::function<void(SceneGenerationContext&)> verifyStep;
-};
-
 void clearCacheDb() {
   std::shared_ptr<CesiumAsync::ICacheDatabase> pCacheDatabase =
       getCacheDatabase();
@@ -224,70 +200,6 @@ bool RunLoadTest(
   return true;
 }
 
-void refreshTilesets(SceneGenerationContext& context) {
-  gLoadTestContext.playContext.refreshTilesets();
-}
-
-bool FCesiumLoadTestDenver::RunTest(const FString& Parameters) {
-
-  std::vector<TestPass> testPasses;
-  testPasses.push_back(TestPass{"Cold Cache", nullptr, nullptr});
-  testPasses.push_back(TestPass{"Warm Cache", refreshTilesets, nullptr});
-
-  return RunLoadTest(GetTestName(), setupForDenver, testPasses);
-}
-
-bool FCesiumLoadTestGoogleplex::RunTest(const FString& Parameters) {
-
-  std::vector<TestPass> testPasses;
-  testPasses.push_back(TestPass{"Cold Cache", nullptr, nullptr});
-  testPasses.push_back(TestPass{"Warm Cache", refreshTilesets, nullptr});
-
-  return RunLoadTest(GetTestName(), setupForGoogleTiles, testPasses);
-}
-
-bool FCesiumLoadTestMontrealPointCloud::RunTest(const FString& Parameters) {
-
-  auto adjustCamera = [this](SceneGenerationContext& context) {
-    // Zoom way out
-    context.startPosition = FVector(0, 0, 7240000.0);
-    context.startRotation = FRotator(-90.0, 0.0, 0.0);
-    context.syncWorldPlayerCamera();
-
-    context.pawn->SetActorLocation(context.startPosition);
-  };
-
-  auto verifyVisibleTiles = [this](SceneGenerationContext& context) {
-    Cesium3DTilesSelection::Tileset* pTileset =
-        context.tilesets[0]->GetTileset();
-    if (TestNotNull("Tileset", pTileset)) {
-      int visibleTiles = 0;
-      pTileset->forEachLoadedTile([&](Cesium3DTilesSelection::Tile& tile) {
-        if (tile.getState() != Cesium3DTilesSelection::TileLoadState::Done)
-          return;
-        const Cesium3DTilesSelection::TileContent& content = tile.getContent();
-        const Cesium3DTilesSelection::TileRenderContent* pRenderContent =
-            content.getRenderContent();
-        if (!pRenderContent) {
-          return;
-        }
-
-        UCesiumGltfComponent* Gltf = static_cast<UCesiumGltfComponent*>(
-            pRenderContent->getRenderResources());
-        if (Gltf && Gltf->IsVisible()) {
-          ++visibleTiles;
-        }
-      });
-
-      TestEqual("visibleTiles", visibleTiles, 1);
-    }
-  };
-
-  std::vector<TestPass> testPasses;
-  testPasses.push_back(TestPass{"Cold Cache", nullptr, nullptr});
-  testPasses.push_back(TestPass{"Adjust", adjustCamera, verifyVisibleTiles});
-
-  return RunLoadTest(GetTestName(), setupForMontrealPointCloud, testPasses);
-}
+}; // namespace Cesium
 
 #endif
